Separate and identify tests runners
